Move GetNodeIPs to pkg/proxy/util
